🔧 move around database structure, remove users
